ragu_acceleration: move MSM equivalence test
